+++ creation_test.v
@@ -1,7 +1,7 @@
 module vtl
 
 fn test_new() {
-	t := new_tensor(shape: [3], init: f64(1.0))
+	t := new_tensor(shape: [3], init: 1.0)
 	varray := tensor_to_varray<f64>(t)
 	assert varray.len == 3
 	assert varray[0] == 1.0
